externalize contactActions
