+++ src/stateful_components/NotificationsPage/NotificationsPage.js
@@ -4,7 +4,7 @@ import styles from './NotificationsPage.module.css';
 import { Notification } from '../../functional_components/Notification';
 import Container from 'react-bootstrap/Container';
 import { connect } from 'react-redux';
-import { fetchNotifications } from '../../actions/UserActions';
+import { fetchNotifications } from '../../actions/userActions';
 import fileUtils from '../../utils/fileUtils';
 const ns = require('solid-namespace')(rdf);
